+++ packages/valory/skills/task_submission_abci/fsm_specification.yaml
